code 格式化，（fmt doc clippy）

+++ src/_1528_restore_string.rs
@@ -2,7 +2,7 @@ pub struct Solution;
 
 impl Solution {
     pub fn restore_string(s: String, indices: Vec<i32>) -> String {
-        // 函数成立 必须 相等 
+        // 函数成立 必须 相等
         if indices.len() != s.len() {
             return "".to_string();
         }
@@ -17,11 +17,10 @@ impl Solution {
 
             // 遍历 indices，将s 中的字符 替换 到res中
             for idx in 0..indices.len() {
-
                 // vec![4, 5, 6, 7, 0, 2, 1, 3]
-                let indx = indices[idx as usize] as usize;
+                let indx = indices[idx] as usize;
                 // c o d e l e e t
-                // res[4] = sd[0] = c 
+                // res[4] = sd[0] = c
                 // sd 索引 0 指向字节 c（原始指针s）再把 c 赋值给 res[4] 的值
                 //（这个前提s, indices 为一一对应的）
                 // 4 5 6 7 0 2 1 3
@@ -29,7 +28,7 @@ impl Solution {
                 // l e e t c o d e
                 // 0 1 2 3 4 5 6 7
                 // 所以这个等式才成立
-                r[indx] = sd[idx as usize];
+                r[indx] = sd[idx];
             }
         }
